Revert "Upgrade to Sidekiq 7 (w/ working feature toggles) (#16047)"
This reverts commit c66ddd0223f08b129bf5bd22cea15b0a61be805a.
