+++ Gemfile
@@ -174,7 +174,7 @@ end
 
 group :test do
   gem 'apivore', git: 'https://github.com/department-of-veterans-affairs/apivore', tag: 'v2.0.0.vsp'
-  gem 'mock_redis'
+  gem 'fakeredis'
   gem 'pdf-inspector'
   gem 'rspec_junit_formatter'
   gem 'rspec-retry'
@@ -215,7 +215,7 @@ group :development, :test do
   gem 'rubocop-rails'
   gem 'rubocop-rspec'
   gem 'rubocop-thread_safety'
-  gem 'sidekiq', '~> 7.2.0'
+  gem 'sidekiq', '>= 6.4.0'
   gem 'timecop'
   gem 'webmock'
   gem 'yard'
