+++ config/initializers/sidekiq.rb
@@ -59,8 +59,6 @@ Rails.application.reloader.to_prepare do
     end
 
     # Remove the default error handler
-    config.error_handlers.delete(Sidekiq::Config::ERROR_HANDLER)
+    config.error_handlers.delete_if { |handler| handler.is_a?(Sidekiq::ExceptionHandler::Logger) }
   end
-
-  Sidekiq.strict_args!(false)
 end
