+++ config/redis.yml
@@ -197,9 +197,6 @@ test:
   <<: *defaults
   redis:
     inherit_socket: true
-    url: <%= Settings.redis.app_data.url %>
-  sidekiq:
-    url: <%= Settings.redis.sidekiq.url %>
 
 production:
   <<: *defaults
