+++ lib/sidekiq/semantic_logging.rb
@@ -4,11 +4,6 @@ require 'sidekiq'
 require 'sidekiq/job_logger'
 
 class Sidekiq::SemanticLogging < Sidekiq::JobLogger
-  def initialize
-    logger = Rails.logger
-    super(logger)
-  end
-
   def call(_worker, item, queue)
     logger_tags = {
       class: item['class'],
