+++ modules/ask_va_api/app/services/crm/cache_data.rb
@@ -4,7 +4,7 @@ module Crm
   class CacheData
     attr_reader :cache_client, :service
 
-    def initialize(service: Service.new(icn: nil), cache_client: AskVAApi::RedisClient.new)
+    def initialize(service: Service.new(icn: nil), cache_client: RedisClient.new)
       @cache_client = cache_client
       @service = service
     end
