+++ spec/lib/chip/redis_client_spec.rb
@@ -29,16 +29,10 @@ describe Chip::RedisClient do
     let(:token) { 'test_token' }
 
     it 'saves entry' do
-      expect_any_instance_of(Redis::Namespace).to receive(:set).once.with(
-        tenant_id, 'test_token', { ex: REDIS_CONFIG[:chip][:each_ttl] }
+      expect_any_instance_of(Redis).to receive(:set).once.with(
+        "chip:#{tenant_id}", 'test_token', { ex: REDIS_CONFIG[:chip][:each_ttl] }
       )
       redis_client.save(token:)
-    end
-
-    it 'saves entry with chip namespace' do
-      redis_client.save(token:)
-
-      expect(redis_client.redis_namespace.redis.keys).to include("chip:#{tenant_id}")
     end
   end
